Starting refactoring the application.
Moved js files into app/
Created a new module named range.js
Added unit tests for range.js on tests/range.test.js
Improved package.json a little bit to use node_module's path.

IN PROGRESS
Having difficulties importing range.js using browserify. I can see the file being added to release.js, but I keep getting Range is undefined when executing the application.
